Snapshot nguon JX1 VS2022 truoc khi dung lai he bang hoi JX2

+++ Sources/Core/Src/KBuySell.h
@@ -43,7 +43,7 @@ public:
 #ifdef _SERVER
 	void			OpenSale(int nPlayerIdx, int nShop, int nShopMoneyUnit);
 	void			OpenSale(int nPlayerIdx, int nSaleType, int nMoneyUnit, int nShopNum, int *nShopId);
-	BOOL			Buy(int nPlayerIdx, int nBuy, int nBuyIdx, BYTE nBuyNumber);	// 买第几个买卖列表中的第几项道具
+	BOOL			Buy(int nPlayerIdx, int nBuy, int nBuyIdx, int nBuyNumber);	// 买第几个买卖列表中的第几项道具
 	BOOL			Sell(int nPlayerIdx, int nBuy, int nIdx, int nBuyNumber);
 	BOOL			AutoSell(int nPlayerIdx, int nBuy, int nIdx, int nBuyNumber);
 	BOOL	CanBuy(int nPlayerIdx, int nBuy, int nBuyIdx, int nBuyNumber);
